+++ src/data.js
@@ -3,6 +3,7 @@ D.settings = {
     cash: {display(v){ if (in_prest) { calc_mana(); } return format_num(v, 1)}, round: 1, do_w: false},
     mana: {display(v){ if (in_prest) { calc_mana(); } return format_num(v, 1)}, round: 1},
     prest_cost: {display(v){ return format_num(v, 1)}, round: 1},
+    calc_mana: {display(v){ return format_num(v, 1)}, round: 1},
     idle_orb_sec: {display(v){ return format_num(v, 1)}, round: 1},
     idle_orb_cost: {display(v){ return format_num(v, 1)}, round: 1},
     burst_fire_amount_cost: {display(v){ return format_num(v, 1)}, round: 1},
@@ -15,6 +16,8 @@ const calc_mana = ()=>{
     return D.calc_mana;
 }
 
+D.debug = [0, 0, 0, 0];
+
 D.$load = true;
 D.mana = 0;
 D.calc_mana = 0; 
@@ -22,11 +25,14 @@ D.prest_cost = 100000; // 100k
 D.prest_upgr_costs = [5, 5, 25, 10, 10];
 D.prest_upgr_values = [0, 0, 2, 0, 0];
 
+D.draw_setting = 0;
+
 D.$init = {};
 D.$do_init = true;
 D.cash = 0;
 
 D.ent_len = entities.length;
+
 
 //#region [> Orb Stuff <]
 //* Clicking
@@ -44,7 +50,7 @@ class Orb {
 const orb_colors = ["#aaa", "#00ee33", "#9999ff", "#dd55dd", "#ff722b", "#ffc23d"];
 D.orbs = [
     // (    color,         weight,   value,   upgrade cost,   per upgrade value)
-    new Orb(orb_colors[0], 1,        0.1,     5,              0.1, 1),
+    new Orb(orb_colors[0], 1,        0.1,     5,              0.1),
     new Orb(orb_colors[1], 0,        1,       50,             1),
     new Orb(orb_colors[2], 0,        5,       500,            5),
     new Orb(orb_colors[3], 0,        50,      40000,          50),
@@ -70,4 +76,4 @@ window.onbeforeunload = ()=> local.store();
 // })();
 
 
-const F_acos = (x) => { return (-0.69813170079773212 * x * x - 0.87266462599716477) * x + 1.5707963267948966; }
+const F_acos = (x) => { return (-0.698 * x * x - 0.872) * x + 1.570; }